LayoutTests/imported/w3c/web-platform-tests/css/css-writing-modes/padding-percent-orthogonal-dynamic-expected.txt


FAIL 400px minus 100px padding based on #child height assert_equals: expected 300 but got 800
FAIL 200px minus 100px padding based on #child height assert_equals: expected 100 but got 800
